Bump arches core version, move submit project into subfolder, remove arches import workaround, add querysets dependency, add endpoint to get blank heritage site tree, use alias not relative paths
