+++ bcrhp/media/css/project.css
@@ -1,5 +1,3 @@
-@import "tailwindcss";
-
 .header-link {
     color: white;
     font-size: 0.8rem;
