+++ bcrhp/templates/arches_urls.htm
@@ -9,7 +9,9 @@
 <div class="arches-urls"
      api_user="{% url 'api_user' %}"
      user_profile_manager="{% url 'user_profile_manager' %}"
-     dashboard="{% url 'submissions' %}">
+     dashboard="{% url 'submissions' %}"
+     api_resource_blank='(graph) => {return "{% url "arches_querysets:api-resource-blank" "----p1" %}".replace("----p1", graph)}'
+     api_resource='(graph,pk) => {return "{% url "arches_querysets:api-resource" "----p1" "00000000-0000-0000-0000-000000000000" %}".replace("----p1", graph).replace("00000000-0000-0000-0000-000000000000", pk)}'>
 </div>
 
 {% endblock arches_urls %}
