+++ package.json
@@ -24,16 +24,21 @@
         "moment-timezone": "^0.5.45",
         "nomnom": "npm:@gerhobbelt/nomnom",
         "rimraf": "^5.0.7",
+        "@turf/helpers": "7.2.0",
+        "@turf/centroid": "7.2.0",
+        "@turf/bbox": "7.2.0",
         "underscore": "^1.13.6"
     },
     "devDependencies": {
         "@eslint/js": "^9.20.0",
+        "@html-eslint/eslint-plugin": "^0.41.0",
+        "@html-eslint/parser": "^0.41.0",
         "@primevue/auto-import-resolver": "^4.3.1",
         "@typescript-eslint/parser": "^8.24.1",
         "@vue/eslint-config-prettier": "^10.2.0",
         "@vue/eslint-config-typescript": "^14.4.0",
         "arches-dev-dependencies": "archesproject/arches-dev-dependencies#stable/7.6.12",
-        "eslint": "^9.20.1",
+        "eslint": "^9.29.0",
         "eslint-plugin-vue": "^9.32.0",
         "globals": "^16.0.0",
         "lightningcss": "^1.29.2",
@@ -42,33 +47,31 @@
         "typescript": "^5.6.2",
         "typescript-eslint": "^8.24.1",
         "unplugin-vue-components": "^28.4.1",
+        "vite-plugin-css-injected-by-js": "^3.5.2",
+        "vite-plugin-node-polyfills": "^0.23.0",
         "vue-eslint-parser": "^9.4.3"
     },
     "dependencies": {
         "@ckeditor/ckeditor5-vue": "7.3.0",
-        "@primeuix/themes": "^1.0.0",
-        "@primevue/forms": "^4.3.2",
-        "@tailwindcss/vite": "^4.0.13",
+        "@primeuix/themes": "^1.1.1",
+        "@primevue/forms": "^4.3.5",
         "@vitejs/plugin-vue": "^5.2.1",
         "@vue/cli": "^5.0.8",
-        "arches": "github:bcgov/arches#7.6.12+bcgov.1",
+        "arches": "github:bcgov/arches#v7.6.12.1_bcgov",
         "arches-component-lab": "github:archesproject/arches-component-lab#main",
-        "bcgov_arches_common": "github:bcgov/bcgov-arches-common#release/1.2.7",
+        "arches-querysets": "github:archesproject/arches-querysets#main",
+        "bcgov_arches_common": "github:bcgov/bcgov-arches-common#brf/feat/add_goeom_dropzone",
         "ckeditor5": "44.2.0",
         "dayjs": "^1.11.13",
         "dompurify": "^3.3.0",
-        "js-cookie": "^2.1.1",
-        "primevue": "^4.2.5",
+        "js-cookie": "^3.0.5",
+        "maplibre-gl": "^5.10.0",
+        "primevue": "^4.3.4",
         "quill": "^2.0.3",
-        "rollup": "^4.34.8",
-        "tailwindcss": "^4.0.13",
-        "tailwindcss-primeui": "^0.5.1",
         "vite": "^6.2.1",
-        "vite-plugin-css-injected-by-js": "^3.5.2",
-        "vite-plugin-node-polyfills": "^0.23.0",
-        "vue": "^3.5.13",
-        "vue-router": "^4.4.3",
-        "zod": "^3.23.8"
+        "vue": "^3.5.16",
+        "vue-router": "^4.5.1",
+        "zod": "^4.1.5"
     },
     "nodeModulesPaths": {}
 }
